+++ CHANGES.md
@@ -5,9 +5,13 @@ E-Maj - Change log
 ###Enhancements:###
   * Adjust the behaviour of some functions that process past time frame,
     when tables have been removed from their group. Now
-    emaj_delete_mark_group(), emaj_get_consolidable_rollbacks() and 
-    emaj_consolidate_rollback_group() functions take into account the real
-    group content on the time frame they process.
+    emaj_delete_mark_group(), emaj_get_consolidable_rollbacks(), 
+    emaj_consolidate_rollback_group() and emaj_snap_log_group() functions take
+    into account the real group content on the time frame they process.
+  * The emaj_snap_log_group() function now returns the number of generated
+    files, and the generated file names are directly derived from log table
+    names (so E-Maj names prefix defined into the emaj_group_def table are now
+    used to build the output file names).
   * Various minor code changes.
 
 ###Bug fixes:###